pyproject.toml
[project]
name = "ai-framework-examples"
version = "0.1.0"
description = "AI Framework Examples - CrewAI and Pydantic AI task implementations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "pydantic-ai>=0.0.14",
    "pyyaml>=6.0",
    "boto3>=1.34.0",
    "arize-phoenix>=4.0.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "openinference-instrumentation-pydantic>=0.1.0",
    "logfire>=0.40.0",
    "pytz>=2024.1",
    "langchain-community>=0.3.0",
    "sqlalchemy>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
